Fifo implementation
