+++ backend/src/schedule/schedule_builder.rs
@@ -1,0 +1,4 @@
+pub struct DayBuilder {
+    slots: [Slot; 5],
+    money: u64,
+}
